Fix issues on SigComp

+++ branches/2.0/doubango/tinySIGCOMP/src/tcomp_compartment.h
@@ -55,8 +55,8 @@ typedef struct tcomp_compartment_s
 	tcomp_states_L_t *local_states;
 	tcomp_params_t *remote_parameters;
 	tcomp_params_t *local_parameters;
-	uint16_t total_memory_size;
-	uint16_t total_memory_left;
+	uint32_t total_memory_size;
+	uint32_t total_memory_left;
 
 	tcomp_buffer_handle_t *lpReqFeedback;
 	tcomp_buffer_handle_t *lpRetFeedback;
@@ -70,7 +70,7 @@ typedef struct tcomp_compartment_s
 }
 tcomp_compartment_t;
 
-tcomp_compartment_t* tcomp_compartment_create(uint64_t id, uint16_t sigCompParameters);
+tcomp_compartment_t* tcomp_compartment_create(uint64_t id, uint32_t sigCompParameters);
 
 //
 //	SigComp Parameters
@@ -87,7 +87,7 @@ void tcomp_compartment_freeStateByPriority(tcomp_compartment_t *compartment);
 void tcomp_compartment_freeState(tcomp_compartment_t *compartment, tcomp_state_t **lpState);
 void tcomp_compartment_freeStates(tcomp_compartment_t *compartment, tcomp_tempstate_to_free_t **tempStates, uint8_t size);
 void tcomp_compartment_addState(tcomp_compartment_t *compartment, tcomp_state_t **lpState);
-uint16_t tcomp_compartment_findState(tcomp_compartment_t *compartment, const tcomp_buffer_handle_t *partial_identifier, tcomp_state_t **lpState);
+uint32_t tcomp_compartment_findState(tcomp_compartment_t *compartment, const tcomp_buffer_handle_t *partial_identifier, tcomp_state_t **lpState);
 void tcomp_compartment_freeGhostState(tcomp_compartment_t *compartment);
